--- bin/.env.sample
@@ -1,16 +1,0 @@
-export API_ADDR=0.0.0.0
-export API_PORT=3000
-
-export FRONT_ADDR=0.0.0.0
-export FRONT_PORT=3001
-
-export DATABASE_USER=DATABASE_USER  # Please CHANGE ME (REQUIRE)
-export DATABASE_PASS=DATABASE_PASS  # Please CHANGE ME (REQUIRE)
-export DATABASE_HOST=DATABASE_HOST  # Please CHANGE ME (REQUIRE)
-export DATABASE=DATABASE  # Please CHANGE ME (REQUIRE)
-export DATABASE_URL=postgres://${DATABASE_USER}:${DATABASE_PASS}@${DATABASE_HOST}:5432/${DATABASE}
-
-export FRONT_CSRF_USE=true
-export MCIAM_USE=true
-
-export SESSION_SECRET=SESSION_SECRET # Please CHANGE ME (REQUIRE)
