+++ conf/.env.sample
@@ -12,4 +12,3 @@ export DATABASE_URL=postgres://${DATABASE_USER}:${DATABASE_PASS}@${DATABASE_HOST
 
 export MCIAM_USE=true
 export MCIAM_TICKET_USE=false
-export FRONT_CSRF_USE=false
